==> datasets/classification/10_PrepareDataSetsXFolds.ipynb <==
### Preparing CFV splits for
 Retrieve available data on the disk and validate it
* Match data table with it
* Write file names to the table
* Split table into N-cross splits + separate split for test
* Use table as CSV Dataset to retrieve the image data

import os
import glob
import sys; sys.path.append("../")
import pandas as pd
import numpy as np

from sklearn.model_selection import StratifiedGroupKFold
from sklearn.model_selection import KFold

nfolds = 5
test_split = True

dataset_source = "/home/oleksii/projects/ohif-orthanc-postgres-docker/datasets/classification/prostate_class_dataset_demography_final_psa_vol_20240919_feat.csv"
df = pd.read_csv(dataset_source, sep=';')

df.csPCa

def get_dir_and_basename(path):

    # Split the path into directory and file name
    directory, file_name = os.path.split(path)
    # Get the last directory name
    last_directory = os.path.basename(directory)
    # Combine the last directory name and the file name
    result = os.path.join(last_directory, file_name)
    
    return result 

def get_alta_dataset(data_path):
    studies = os.listdir(data_path)
    records = []
    for study_id in studies:
        
        study_path = os.path.join(data_path, study_id)
        t2ws = glob.glob(os.path.join(study_path, "*tra_t2w.nrrd"))
        adcs = glob.glob(os.path.join(study_path, "*tra_adc.nrrd"))
        dwis = glob.glob(os.path.join(study_path, "*tra_dwi_calc_bval*.nrrd"))
        segs = glob.glob(os.path.join(study_path, "*.seg.nrrd"))
        
        d = {}
        d['study_orthanc_id'] = study_id
        d["t2w"] = get_dir_and_basename(t2ws[0]) if t2ws else None  
        d["adc"] = get_dir_and_basename(adcs[0]) if adcs else None  
        d["seg"] = get_dir_and_basename(segs[0]) if segs else None  
        d["dwis"] = ','.join([get_dir_and_basename(dwi) for dwi in dwis]) if dwis else None  
        records.append(d)
         
    ds = pd.DataFrame(records)
    print(len(ds))
    return ds


data_path = "/data/oleksii/Prostate-Classification-Datasets-NRRDS/ALTA-Classification-Dataset-orient-preprocess/"
dataset_alta = get_alta_dataset(data_path)
print(len(dataset_alta))

# keep full cases
dataset_alta = dataset_alta[(~dataset_alta['dwis'].isnull()) & (~dataset_alta['t2w'].isnull()) & (~dataset_alta['adc'].isnull()) & (~dataset_alta['seg'].isnull())]
# keep only prostate cases
dataset_alta = dataset_alta[dataset_alta['dwis'].apply(lambda x: True if ".nrrd" in x.split(",")[0] else False)]

dataset_alta

# merge preprocessed data with tabular one 
dataset_alta_m = pd.merge(dataset_alta, df, how="inner", on="study_orthanc_id")

dataset_alta_m[dataset_alta_m['csPCa'].isna()]

# add another "multi external dataset"
external_df = pd.read_csv("/home/oleksii/projects/ohif-orthanc-postgres-docker/sequence_mapping/external_rare_cases.csv", sep=";")

dataset_test_ext = dataset_alta_m[dataset_alta_m['study_orthanc_id'].isin(external_df['study_orthanc_id'])]
dataset_train = dataset_alta_m[~dataset_alta_m['study_orthanc_id'].isin(external_df['study_orthanc_id'])]

# test is defined based on cases that have PI-RADS score
test_pirads = pd.read_csv("/home/oleksii/projects/ohif-orthanc-postgres-docker/datasets/classification/prostate_test_set_pirads.csv", sep=";")

dataset_test = dataset_train[dataset_train['study_orthanc_id'].isin(test_pirads['study_orthanc_id'])]
dataset_train = dataset_train[~dataset_train['study_orthanc_id'].isin(test_pirads['study_orthanc_id'])]

len(dataset_test) + len(dataset_test_ext) + len(dataset_train)

print("test_ext:", len(dataset_test_ext))
print("test:", len(dataset_test))
print("train:", len(dataset_train))

# first, get the test set
# sgkfold = StratifiedGroupKFold(n_splits=nfolds + 1, shuffle=True, random_state=42)
# splits = list(sgkfold.split(X=dataset_alta_m, y=dataset_alta_m['csPCa'].to_list(), groups=dataset_alta_m['PatientID'].to_list()))

# # split test from train! 
# dataset_test = dataset_alta_m.iloc[splits[0][1]]
# dataset_train = dataset_alta_m.iloc[splits[0][0]]

# next, get the cross fold set
sgkfold_train = StratifiedGroupKFold(n_splits=nfolds , shuffle=True, random_state=42)
splits_folds = list(sgkfold_train.split(X=dataset_train, y=dataset_train['csPCa'].to_list(), groups=dataset_train['PatientID'].to_list()))

splits_path = "/home/oleksii/projects/ohif-orthanc-postgres-docker/datasets/classification/cross_fold_splits"
os.makedirs(splits_path, exist_ok=True)
df.to_csv(os.path.join(splits_path, os.path.basename(dataset_source)), sep=';', index=False) 

dataset_test.to_csv(os.path.join(splits_path, 'test.csv'), sep=';', index=False)
dataset_test_ext.to_csv(os.path.join(splits_path, 'test_ext.csv'), sep=';', index=False)
for i, split in enumerate(splits_folds):
    print(f"Saving fold {i}")
    
    fold_path = os.path.join(splits_path, f'fold_{i}')
    os.makedirs(fold_path, exist_ok=True)
    
    train_fold = dataset_train.iloc[split[0]]
    valid_fold = dataset_train.iloc[split[1]]
    train_fold.to_csv(os.path.join(fold_path, 'train.csv'), sep=';', index=False)
    valid_fold.to_csv(os.path.join(fold_path, 'valid.csv'), sep=';', index=False)
