==> datasets/classification/11_PIRADS.ipynb <==
Here, a real PI-RADS perfomance of ALTA Klinik is evaluated against biosies

import pandas as pd
from sklearn.metrics import confusion_matrix
from sklearn.metrics import precision_score, recall_score, accuracy_score

pirads_df = pd.read_excel("PI-RADS_bis_26_01_2024.xlsx")
pirads_df.rename(columns={"Pat-ID": "ALTAPatientID", 
                          "Befunddatum": "StudyDate"}, inplace=True)

class_df = pd.read_csv("prostate_class_dataset_demography_final_psa_vol_20240919_feat.csv", sep=";")
class_df = class_df[["ALTAPatientID", "StudyDate", "study_orthanc_id", "csPCa"]]

class_df

pirads_df

pirads_df['StudyDate'] = pd.to_datetime(pirads_df['StudyDate'])
class_df['StudyDate'] = pd.to_datetime(class_df['StudyDate'])

# Merge the tables on PatientID
merged_df = pd.merge(pirads_df, class_df, on='ALTAPatientID', suffixes=('_pirads', '_mri'))

# Calculate the absolute difference in days between the dates
merged_df['date_diff'] = abs((merged_df['StudyDate_pirads'] - merged_df['StudyDate_mri']).dt.days)

# Filter the merged DataFrame to include only rows where the date difference is within 30 days
merged_df = merged_df[merged_df['date_diff'] <= 30]

# Group by PatientID and keep the row with the smallest date difference
closest_studies = merged_df.loc[merged_df.groupby('ALTAPatientID')['date_diff'].idxmin()]


pirads_df_histo = closest_studies.reset_index()

pirads_df_histo

# pirads_df_histo = pd.merge(class_df, pirads_df, on=['ALTAPatientID', 'StudyDate'], how='right')
# pirads_df_histo 


pirads_df_histo = pirads_df_histo[~pirads_df_histo["csPCa"].isna()]

pirads_df_histo['PI-RADS_gt_3'] = pirads_df_histo['PI-RADS'] > 3
pirads_df_histo['PI-RADS_gt_4'] = pirads_df_histo['PI-RADS'] > 4

pirads_df_histo[pirads_df_histo['PI-RADS_gt_3'].isna()]

print("Number of csPCa cases:")
pirads_df_histo['csPCa'].value_counts()



# Determine the minimum number of cases among the classes
min_count = pirads_df_histo['csPCa'].value_counts().min()

# Randomly sample this number of cases from each class
balanced_df = pirads_df_histo.groupby('csPCa').apply(lambda x: x.sample(350, random_state=43)).reset_index(drop=True)

# Display the balanced DataFrame
print(balanced_df)
balanced_df['csPCa'].value_counts()

# note that this is not a complete list of pirads scores!
balanced_df.to_csv("prostate_test_set_pirads.csv", sep=";")



def caluculate_metrics(actual, predicted):# Calculate precision and recall
    precision = precision_score(actual, predicted)
    recall = recall_score(actual, predicted)
    acc = accuracy_score(actual, predicted)
    cm = confusion_matrix(actual, predicted)

    # Extract true positives, false positives, true negatives, and false negatives from the confusion matrix
    tn, fp, fn, tp = cm.ravel()

    # Calculate true positive rate (recall) and false positive rate
    tpr = tp / (tp + fn)
    fpr = fp / (fp + tn)

    print(f"Recall: {recall:.3f}")
    print(f"Precision: {precision:.3f}")
    print(f"Acc: {acc:.3f}")
    print(f"True Positive Rate (TPR): {tpr:.3f}")
    print(f"False Positive Rate (FPR): {fpr:.3f}")
    print(f"tn, fp, fn, tp: {tn}, {fp}, {fn}, {tp}")
    

    print("\nConfusion Matrix:")
    print(cm)


# Assuming the binary columns are named 'predicted' and 'actual'
predicted = pirads_df_histo['PI-RADS_gt_3'].to_numpy().astype(int)
actual = pirads_df_histo['csPCa'].to_numpy().astype(int)
print(f"PI-RADS>3 -> csPCa ({len(actual)} samples)")
caluculate_metrics(actual, predicted)


print(f"PI-RADS>4 -> csPCa ({len(actual)} samples)")

# Assuming the binary columns are named 'predicted' and 'actual'
predicted = pirads_df_histo['PI-RADS_gt_4'].to_numpy().astype(int)
actual = pirads_df_histo['csPCa'].to_numpy().astype(int)
caluculate_metrics(actual, predicted)
